backend/requirements.txt
Django==5.1.3
djangorestframework==3.15.2
djangorestframework_simplejwt==5.3.1
drf_extra_fields==3.7.0
